Fixes LP #920782 - Malformed request URL
Sometime recently, the tenant ID instead of the tenant Name
was used in Compute URLs, and we were appending the tenant Name
to the management URL returned from Keystone improperly.

This patch grabs the tenant ID from the returned management URl
and appends the tenant ID instead of the tenant Name.

Change-Id: I4d4d23b3722c5c67463592397c5b23c5f3281c30