Don't pass None for any values in post body.
There is unspecifiedness and nonuniformity of the behavior of passing
None values to create server. So don't do it.
The case of the 'meta' keyword argument is ugly and it should be changed
to 'metadata' but the various APIs are inconsistent about this so I didn't
want to mess with that now and risk breaking something.

Change-Id: I3a8a3afbc179a54b524db0e2ed04d1542401f4a7