Add tests for volume attach and detach
This adds a test_attach_detach_volume case that attempts to create a
server, attach a volume and make sure that it hangs around during a
power cycle. It also tests to make sure that once we detach a volume,
it stays that way during a power cycle.

Also adds servers_client methods for start, stop, attach_volume and
detach_volume. Further adds get_partitions() to remote_client.

Note that right now, this doesn't succeed with libvirt. A patch to nova
is in the works.

Change-Id: Id43175fd9c8e8cf8971e77f212d0dc74a20def4d