Fix checks in server listing only lok for an ID
The list server filter details checks were checking
entire mappings against the list of mappings returned
from a call to GET /servers/detail. Unfortunately, the
mappings returned by the GET /servers/detail and the
mapping returned by POST /servers are slightly different,
resulting in the comparison failing. We really only are
checking that the server instance we expect to be in the
returned list of instances is there, so checking for the
server ID in the returned list is fine.

Change-Id: If9680f56a1ea06ffe93393d06b23bd231ccddd01
Fixes: LP Bug 1028465pep8 build for tempest.