test_absolute_limits.py to check limits response
This commit adds a test case that checks if all the absolute limit
values are returned in the REST call to Nova API Extension
"/limits". Also modifies limits_client.py file for JSON and XML
requests so as to make it less verbose. 'get_specific_absolute_limit'
method is added in the limits_client.py file. As method
'get_personality_file_limit' is removed, "test_server_personality.py"
is modified.

Fixes: bug #1025294
Change-Id: Id334cd73645273d7b80b25ba698d03dc5599a211