Add an xml/servers_client.py implementation
This needs to inherit some common bits from the json implementation,
but works for now.

This changes openstack.py::Manager to take an interface argument and
select the appropriate client(s) accordingly. This is ugly, and probably
not the best approach, but for the moment, it avoids the class inflation
that would result from creating subclasses at the Manager _and_ Base*Test
layers.

Change-Id: I7814054eab59bc34258fbbd1df43a4140448f448