Tempest tests to cover live-block-migration
Related to bug 1051881.

Two test cases: one to cover the live block migration on XenServer, and
the other to indicate that the machine status is left in an inconsistent
state in case a non-existing host is specified as target.
To run the live-migration tests with block migration, include:

    [compute]
    live_migration_available = true
    use_block_migration_for_live_migration = true

in your tempest.conf

To run these tests, the hypervisors must support the Storage XenMotion
feature.

Change-Id: I1e6cc903fa573ecd219a08bf4dddd9cc1f4f60df