Handle ImportError's when quantumclient is missing
Updates tempest/manager.py so that it ignore's ImportError
exceptions when quantumclient is not present.

Not all Tempest tests require a quantumclient so it makes sense not
to require it all the time. If a user tries to run quantum specific
tests they will simply get an error when trying to create the
network client instead should quantumclient not be present.

Change-Id: Iea5476eea91e0d92c3b4728f041a3ce72bb5a320