be specific about metadata too long error
Tempest was catching a generic fail on the metadata too
long call, but a very specific error is supposed to be returned,
a 413. In real runs there was a bug in nova which crashed on the
request (fixed in https://bugs.launchpad.net/nova/+bug/1079387).

Fix tempest to only accept a 413 as valid on this request.

Make this an assertRaise test instead of a negative test

Change-Id: Id09394abc863ecd773f56dbc8a965d87ae7dde97