Fix issue with 404 logs on wait for delete.
In some cases the wait on volume deletes will receive a 404
error which results in a log output. However for these tests
this is the expected behavior so the log output isn't needed.

This adds a flag to suppress log output on 404 errors, which
is then used on all wait functions.

Fixes: bug 1084283
Change-Id: Id455a22317ecbdf318e9ad2ea0a76ad9f08ca899