Improve pep8 checks to be similar to those in nova
This adds hacking.py, based on the one in nova, to wrap around
pep8 and enforce more stringent checks on tempest. This adds
the rules that nova is currently enforcing on incoming commits.

The rules that pep8 is ignoring are copied from the pep8 check on
jenkins.

The HACKING file was also updated to indicate the inclusion
of the new rules that are being enforced.

Change-Id: I5c85829ee15d4264d8de2d9f0207d9dd3dfbbff2