Tempest should ignore SSL certificate validation
In testing and QA environments, often endpoints are set up
with self-signed or incomplete certs, and Tempest will bomb
out with complaints about being unable to validate SSL certificates.

Tempest isn't about testing SSL certificates. It's about testing
the OpenStack APIs, and if a common scenario is to test SSL and
endpoints in a testing environment, the client should disable
SSL cert checking.

Change-Id: I999cef9b5cd40a94c1b1bd63f4e19b2e58924c9a
fixes: lp bug #1089762