Logic in rest_client incorrect "resp.status=413"
rest_client is NOT redriving rate limited calls for the second time.
When 413 error code is encountered for a request, it checks for the
word "overLimit" and "limit" in resp_body and resp_body[message]
respectively so as to confirm if it is a absolute-limited call.
If absolute limited call is confirmed, OverLimit exception is raised
without any re-try for request.
Then comes the check for depth < MAX_RECURSION_DEPTH, if it is true,
re-drive towards request call occurs.
But, the word "limit" also appears in the message of ratelimited call.
Hence it is incorrect to raise OverLimit exception without re-try when
word "limit" is present in the message of response.
Thus changed the word to "exceeded" instead of "limit".
Fixes: bug #1093573

Change-Id: Iadd94c18668096d7e4795631337230ec36e9c4b9