fix file injection test to not assume /etc is present
File injection is currently failing as the /etc
directory is not present and the file injection
logic currently doesn't ensure a specified directory is present.
There is a pending change https://review.openstack.org/#/c/18900/
that changes operation so that a failure to inject user
specified files results in startup failure,
so adjust the test to inject to always present root dir.

Change-Id: I864272516e06ac0c2317b4470e263f5db41149f9