Add wait for resource deletion on volume teardown
The teardown routines in some of the tests were not
waiting for verification that the volume was actually deleted.
As a result it's easy to intermittently hit a case where we run
out of space for volume create in the following tests.

This patch just adds a resource_deletion check/wait after
each volume delete operation to make sure we don't get ahead
of ourselves.

Change-Id: I47f9ad2b8e6f79348c3d6188da3c24af0c6fe5d4
Fixes: bug 1092689