Disable test_run_terminate_instance
Looks like the nova API now gives an error response instead of an
empty list. It can be a correct behaviort, but it does it with an incorrect
error code.

This change let us pass on the https://review.openstack.org/#/c/19098/
change and continue the work on tempest and nova side.

Change-Id: I18309429e9967ccd7804ff0cc1c385ece100dd67