Handling rate-limit for JSON request- rest_client
Modified rest_client.py so as to re-try when request is rate-limited
with JSON as interface. The current logic in rest_client.py would not
re-try and directly raise "exceptions.OverLimit" once key 'overLimit' is
present in the response body. It does not see if the overLimit is due
to rate-limit. Hence checking if word "rate-limited" is present in the
response message.

Change-Id: I9dd3e2b4a117e8e905bf3f1ba8d108689e602297
Fixes: bug #1098540