Use '-m' for tempest_coverage.py in tox.ini
Currently you can't run tempest_coverage.py because of an import
failure when trying to import from files in tempest. This is because
the python path from tempest/tools doesn't include it's parent
directory. This commit gets around the import failure with tox by
using 'python -m' for the script.

Change-Id: Iebbbbc1fbcf7f5fb0eb5d9dbdf3901cb77b90a43