Convert these tests of ServerBasicOps into one test.
These tests must run in the fixed order. When tempest run in one
process, it's ok, because these tests run in alphabetical order.
But when tempest run in multiple process, these tests will run in
random order, some tests will fail. So I think we should make them
one test, not a set of tests. And when it fail, it can give different
information in different stage. I think it's ok.

Part of blueprint speed-up-tempest

Change-Id: I993c11050407786823c99dd93fc514d5627eb433