Only create 1 server for server actions tests.
This switches the setUp() and tearDown() methods in
test_server_actions.py to be setUpClass() and tearDownClass().
The server create and delete for each test method were taking
a significant amount of time. This minimizes the time spent for
those operations by only using one server since there wasn't a
need to use separate servers for each test.

Change-Id: I5e7a3831c9f018e871c0fcfcd575e662a143cd9b