Revert "Split up XML and JSON testing."
This reverts commit 58dec9d45360a69192fc6ba43dddf5a597063972

This ended gating of the XML API without any real public deliberation.  It should be reverted, and a plan agreed upon with PTLs and the CI team.

Also, if we do decide to implement this, we should make the CI change _first_ so that we don't abruptly stop gating on something we should be testing.