Proposed EC2 OpenStack extension
'PAUSED' OS instance status is not mappable to any EC2 status correctly.
Adding a 'paused' status seams to the only viable way to handle this.

Normally you cannot reach this status just with the EC2 API, but all other
defined EC2 statuses are schematically incorrect for status mapping.

A suspend can be considered as a 'stopped' status.
'rebooting' can be considered as 'pending' status.

Using the 'paused' status is only valid when the VM paused because
of the user's explicit request, so it is not temporary pause because
of an internal operation sequence. If the nova just pauses the VM
just temporary it must report 'pending' (or 'running' ) status.

Change-Id: Id4500dd6e38e078d0b0b76fb0dd6a81c054ef646