Removes use of nose.tools.raises
Removes uses of nose.tools.raises and replaces it with
assertRaises. This makes it consistent with how expected
exceptions are handled in other tempest tests and is part of the
process of removing dependencies on nose.

For the files that were altered use of attr(type='negative') were
also removed. These are not used as the tests already handle the
expected failures by catching the exception (formerly by using
@nose.tools.raises decorators). These need to removed eventually
as the nose.plugins.atttrib.attr decorators are replaced by the
testtools attr decorators which will remove the final
dependencies on nose (but nosetests will still be able to be
used, it just won't be required).

Change-Id: Id8a86da283fd16dd5a907be28d442c23c17cb35c