Remove duplicate calls to clear_servers
Removes ServerActionsTest* call to the "clear_servers" function
in its class teardown fixture as it is redundant.
Removes ServerActionsBase* call to the "clear_servers" function
in its test teardown fixture as the clean-up is down at
class teardown anyways.
Moves some duplicated code in test_server_actions to a class
function rebuild_servers, which is used at class and test setup
to create a new server when needed.

Fixes: bug #1124259
Change-Id: I29b3a4ca6e8738104452976735f49c331ffa116f