Update stress tests to properly use tempest.config
When tempest.config was switched over to using oslo the stress tests
weren't updated with those changes. Also subsequent changes to the
tempest options were not updated in the stress test suite. This left
the stress tests unusable because it couldn't get the needed
information from config. This commit updates the stress test suite
to properly use tempest.config. It also adds a new section to the
config file for stress test specific options.

Fixes bug 1133012

Change-Id: I1b0c726f53dc0e3e0777e27af3e860d35029f958