Use oslo.config-1.1.0b1
The cfg API is now available via the oslo-config library, so switch to
it and remove the copied-and-pasted version.

Add the oslo.config-1.1.0b1 tarball to tools/pip-requires - this will
be changed to 'oslo-config>=2013.1' when oslo-config is published to
pypi. This will happen in time for grizzly final.

Remove the 'deps = pep8==1.3.3' from tox.ini as it means all the other
deps get installed with easy_install which can't install oslo-config
from the URL.

Change-Id: Ie803935d5f1b4fb88e76932cc42ce22b45acffad