Update defaults for s3 materials paths/names
devstack has upgraded to downloading the cirros-0.3.1 image.
This change updates the ari/ami/aki manifest defaults in
tempest.conf.sample to match as otherwise the relevant tests
are fairly quietly skipped.

Adds a configuration option to set the timeout for waiting on
remote ssh commands to finish. This is necessary as the
test_integration_1 now executes with this fix and times out
on the CI systems due to the ssh command taking longer than
the hardcoded timeout.

Change-Id: I744d242416ed63ab49facae70948d5c007df13c7