More assertions for test_integration_1
For the boto/test_ec2_instance_run.py test_integration_1 test
add assertions to check that the security group modifications
and public ip association calls actually succeed.

This is to help debug the intermittent bug 1117555 which appears
to be due to the instance sometimes being unreachable via its
public ip.

Change-Id: I8c1ffeee62e1021ecf24f05ac26518ced640c5a1