enable test_servers_negative
this was disabled when we were seeing flakey fails during Folsom
release, however all but one of these tests current pass. And the
failure is a new nova bug. Pushing to CI to verify that bug on both
databases before providing the fixes to nova (of which there will
be at least 2)

fix xml clients for negative tests

the xml clients were failing additional tests because the clients
didn't support all the features as the json tests. This enables
security_groups and networks in the xml client for create.

It also skips the test_create_numeric_server_name as that can't be
tested in xml, it always turns it into a string.

Fixes bug #1046870

Change-Id: I8a3a0bef9b1d134da369242a0cf14c3dcb61c6de