Fix endpoint usage for glance_http in image client.
Previously the image client was using python-keystoneclient to
generate a new token and to get the endpoint for glance to use
with the glance http library. This was causing permission issues
if the rest client was used to try and delete an image created using
glance_http.

This commit corrects the behavior so that the token and endpoint are
acquired the same way as the rest client so there is no potential
permission issues between the http libraries.

Change-Id: Idff5fcb82019c6b807b87dda480fbcf0b6f8aef2