Have paramico to register the event pipe in time
paramiko is multi-thread package which has a dedicated thread for input
processing.
paramiko using a pipe trick to communicate his buffered_pipe state for
event based wait operations.
It is possible the reading (transport thread) gets an EOF (or CLOSE) message before
the event pipe created, unfourtunatly it leads to the EOF and CLOSE
messages not communicated correctly by the event pipe which is created later.

The applied change expected to be compatible with newer paramiko
versions, basically it haves the paramiko to create the pipe earlier.

test_integration_1 test case enabled.
Fixing bug ##1117555.

Change-Id: I5e7bff253ed21fc424acc426545cc48f22e96d74