Fix test_flavor_extra_specs
This commit makes a number of changes to test_flavors_extra_specs
that cleans up a few of the tests in the file. Firstly, the
test_flavor_non_admin_set_get_unset_keys is split up into 3
separate tests. The original method, besides not working, was
testing 3 different permission cases. Secondly, an unused import
was removed. Thirdly, attrs were added for the negative tests in
the file. Lastly, all the try except blocks for the negative
tests were removed.

Change-Id: I4beba91918fc795628f2668b196a3bb150a610d8