raise the first exception in flavors and security_group test
The exceptions raised in the finally block hides the exceptions raised
in the try block.

For example if a resource creation fails you would see the exception from the
resource deletion.

This change replaces the finally usage by addCleanup methods.

test_security_group_rules_create_delete removed because of a duplication.

Change-Id: Id69a3b7edd0933f4d5967bd8b84fb16a2095ab0a