Add api version detection support to glance tests.
This commit adds support for detecting the api versions supported
by glance. The list of versions is aquired by making a GET
request on the glance endpoint.

A new get_versions method is added to rest_client which will make
a GET call on the base_url of the RestClient object. The method
parses the response from the api call to get the list of
versions. A check is then added to the base test classes to ensure
the api version being tested is in the list of supported api
versions.

Change-Id: If60bb509bc1b2c7ea514066d12e0a03a41308368