Add duplicate bug detection to skip_tracker.py
Previously, if a tempest test was skipped because of a bug that
was marked a duplicate of another bug the skip tracker would not
check the status of the duplicate bug. This would cause fixed bugs
from being shown up in the list, because the original bug's status
is not necessarily updated if it is marked as a duplicate.

This commit fixes this behavior by checking if there is a duplicate
bug first. Then after all the bugs' statuses are checked the list of
duplicate bugs is checked. If a duplicate bug is marked as fixed the
original bug is added to list of bugs that is safe to unskip.

Change-Id: Ia1aaec9e01a97556d17b0640a95e03f0f0b37680