Remove unnecessary asserts from test_images.py
This commit removes unnecessary assertTrue()s from the tests in
CreateRegisterImagesTest test class in
tempest.tests.image.v1.test_images. The asserts were being used to
verify a particular field in the response, however this isn't
needed, because get() will throw an error if the key is invalid.
Additionally, for those types of test assertIn() should be used
instead of assertTrue(), but since they are unnecessary the asserts
are just removed.

Change-Id: I3bd24d48eb9bbc11f39d8077d1e823b10e34c6d9