Make skip_tracker bug keyword detection more robust.
This commit changes the regex for detecting a bug in
skip_tracker.py to be case insensitive and support a colon
instead of whitespace immediately following the bug keyword.
Previously only Bug or bug was detected, and whitespace needed
to follow the bug keyword for skip_tracker.py to pick up the bug
number correctly. This causes issues with some skips because they
didn't comply with these constraints. This commit fixes these issues
by allow any capitalization of 'bug' and allowing either a colon or
whitespace to immediately follow the bug keyword.

Change-Id: I0e9a073b195d475e0b1eebc76e4db160742e171e