Remove skips in volume types tests.
This commit removes the skips from test_volume_types_extra_specs_negative
for bugs that have been marked as fixed. However, the XML client class was
not written to accept the bad input for the negative tests. So the XML
client is modified to accept this input. Additionally, the skip in
test_volume_types_negative.py was removed as well.

Change-Id: I45090c29a4115c205b656777e57a940a7de7dbf5