cleanup resource in setUpClass if ecxeption raised
this catch the exception in setUpClass of test_list_image_filters
and cleanup resource. Every exception should be catched
so hasn't specify what kind of exception after except.
In BaseComputeTest.create_server if a exception raised in
wait_for_server_status, the server won't be append to cls.servers
and cann't be deleted in clear_server.So change the order.

fix bug 1158851

Change-Id: I4c6cf1e9103383dbb5b572b0ee613a12b8c3be03