Reduce chance of name collision for resources.
I am not sure how to make this totally reliable without using long uuids but this should reduce
the chance of collision from one in thousands to one in millions, more than enough to make this not be the weakest link
of build failure. It can be improved in the future if necessary.

Change-Id: I0337db0e1bf7a028855a3d18b8eb9894e1a436bf