attempt to get to flake8/hacking plugins
this is the infrastructure changes, plus 1 fix, to get us towards
flake8 and hacking plugins.

We need to remove an exit call in __init__ for config to get this
to pass. I think long term this gets addressed by config becoming
a test resource, but it will take some time at summit to figure
that out.

Change-Id: Iedd7931e85da5518cb2a8d58717e37b805267d2c