Initial heat orchestration tests.
This does a creat/list/show/delete of a stack with no resources.

It should run fast and is suitable for gating, since it doesn't
create any nova instances.

I'd consider it to be a smoke test. It confirms that heat-api
can communicate with heat-engine, and that heat-engine has
a properly configured database.

Blueprint: add-basic-heat-tests
Change-Id: Id249a34c3a28e1e69d9eb000c6c5c2e82d619178