change test_register_http_image to use explicit url
previously test_register_http_image did some black magic with
sending an image into swift then using swift as an http server
for the purposes of this test. However there is no other requirement
in this test that swift is actually running, so it breaks under
a default devstack env (and possibly other enviroments).

The real test is whether an http image can be pulled into glance.
Use the cirros image as a default, but allow people to specify
another value if they don't have internet access.

attempt to fix pep8

Change-Id: I3800f4bf7174b85aab747f0c7fe911acac033383