proposal for tempest directory restructure
This is the text of a proposal for tempest directory restructure,
largely coming out of OpenStack summit. Once this has gotten core
sign off, I'llstart implementing the refactors into those
directories, and each will have a more specific README explaining
good tests in that kind of environment.

This proposal doesn't attempt to layout structure for tests we
don't have yet, this is entirely about organizing what we have into
something easier to understand. It is assumed new classes of tests
will get proposed over time, and we'll take them as they come.

Part of bp:tempest-repo-restructure

Change-Id: Ib8c55602b085622d322676ba2c24d9719fb9737d