make status_timeout a method
status_timeout was previously a function, which took an object as
the first parameter always. So it was called as:

test.status_timeout(self, ...)

which is craziness. If self is always the first param it should be
a method to make it.

self.status_timeout(...)

This is needed as part of the scenario refactoring.

part of bp:tempest-repo-restructure

Change-Id: I6c9570155e92711f99129e8e8736ddecc50e1f26