move boto tests to thirdparty directory - part 1
this is part one of the boto refactor, there are still some
shared base classes (and the ec2 & s3 clients are added to every
test case still), however this gets the front end of the boto
runs moved to the thirdparty directory.

3rdparty in the proposal had to become thirdparty in implementation
as python doesn't like modules that start with numbers. Oh well.

We continue to run these tests as part of normal runs by adding
them to the tox.ini file.

Part of bp:tempest-repo-restructure

Change-Id: I30df582a3183453f6c7b36b1ea6857195e953950