initial seed for tempest doc directory
This is a first bit of experiment to get a doc tree for tempest
that will publish our doc content to the web as part of a larger
QA doc guide.

Update some of the base READMEs to be a little more accurate,
now that we're going to be pushing them to the web.

Added tox [venv] to ensure that docs get built

Change-Id: Id54e8fbc1d884525087630106e17b37dc1206cc9