Limit tests should pass correct image and flavor
The existing absolute limit tests were passing hard-coded values for
`flavor_ref` and `image_ref`. This patch switches it to use the values defined
in the tempest config.

Since these hard-coded values were incorrect, it imposed an order on which the
validation checks in Nova had to be run, which is not something we want.

By passing correct data, we're only testing that the limit test was performed,
not that it was performed before the image-check, for example.

Change-Id: I93e29c3be0f257de65a8e87b3552616786e205db