Fixes the multi-backend skip bug and the cleanup steps
The multi-backend test class and the tests within should now be
skipped correctly also when running on python 2.6. Also fixes
a cleanup problem occurring in case of error in setUpClass and
speed things up by not creating the 2nd volume/type if backends
have the same name (in which case only part of the tests can be
executed).

Change-Id: Ibeb961ea000bb1d1034f857a64552c7aa75ed4d5
Fixes: bug #1177608