add create_image_from_server to base class for auto cleanup
this test's cleanup logic was fragile enough that it caused additional
breaks when it fell over, leaving resources around. Create a
create_image_from_server as an equiv of create_server to ensure
that images are auto cleaned up at the end of the run, or if we
fail.

Added logging for exception if found during clear_images

Added logging if we failed in setupClass

Fixes bug #1168586

Change-Id: I1cbd904a71eb81926195c3e7894003fe5016b12e