Remove TODO note
Both postgresql and mysql headers are necessary for certain tox jobs,
but both installed by when it is necerray. (pep8 job in a venv)

It was an issue in the past, but it was solved.
We was not able to combine both tools/pip-requires and tools/test-requires.

The pep8 jobs was done by a shell script, the shell script is removed.
tox aims to be the single entry point for all gate jobs.

Change-Id: Iec5066a30b51fb4c91465a23606ad5cd973bc40c