Fix test_volumes_get volumes cleanup.
Currently for test_volumes_get if the volume creation fails and we won't define
the volume variable which will result in an undefined error in the finally
block that always gets run. This will mask the actual cause of the test
failure. This commit fixes this by reworking the cleanup logic in the tests
to use the addCleanup so if the api call fails the actual error will be printed.

Change-Id: I2619575fe80d4f3ce38cc951e7dbd44311d61432