Avoid potential race condition in list_stacks assert.
Asserting that stack listing has increased by exactly 1
after a create is racy if there are other concurrent stack
operations occuring.

This replaces the assert with a check that confirms the created
stack appears in the stack list.

This race has not actually been triggered yet, but may when tempest
gains some concurrency.

Change-Id: I200efa6d3d55e403c9d9a5f3abc49ad6c024e6a9