Create stack in class setup, rather than instance.
The intent of creating the instance in a setup method
was so that the test methods could share a single instance to perform
their tests.

However, doing it in setUp() resulted in an instance per test method,
which is the opposite of the original intent.

Moving instance creation to the class setup means only one instance is
created.

Change-Id: Ie54573e7371e53a60350a3693a740ac56925b520