improve minDisk checking
If a new image is created from an instance, it is expected that the new
image's minDisk is the same as the original images. This is not valid
with at least with VHD images. See the bug description for a link to a
unit test that proves this. This patch checks if the minDisk is either
the original minDisk or the disk size of the flavor.

Fixes bug 1198265

Change-Id: I59ecb79c933e44b4a9f2d565e050a8f9fd9ac767