Rework stress to be more UnitTest like
This patch reworks the stress test framework and
the scripts to be a bit more flexible and to be
like how unit tests are written.  Instead of
writing a module that contains functions, we now
write an object that can implement a setUp and tearDown
method.  By default the object's 'run' method will be
called by the driver.

The setUp method will be called prior to calling run
and tearDown will be called after run is stopped.

Also added debug logging to the cleanup mechanism.
Also added removing snapshots prior to removing
volumes during cleanup

DocImpact

Implements:  blueprint stress-tests

Change-Id: I39b8a1d38f70ddda4867126b58bb7053b45654d6