Ignore the default quota values
The default Quota values are just values from config file,
and they are usually different at ever cloud provider.

In the API compatibility perspective the values are not
important, but the keys presence is important.

The other option would be, adding the expected response to the
tempest config.

The test_get_updated_quotas test case also got a dedicated tenant.

Fixing bug 1130461
Change-Id: I50204dc7955d68491ad63f239c42fb47591a0be7