Fix tox job for stress tests.
This commit fixes the behavior of the tox stress job. To do this a
new json tests file was added for the tox job. This json file will
contain a list of all the actions that will get run as part of the
tox job. (and by extension the periodic jenkins stress test job) The
tox job is then updated to run each of the tests listed in the json
file for an hour each.

Part of: blueprint stress-tests

Change-Id: I5eb0652799601631e34cc2bbef836624d71bb3e7