Add cinder_available config option.
This commit adds a new config group service_available that will be
used to specify whether a particular OpenStack project is expected to
be running or not. This group is added with one member cinder which is
used to specify whether cinder is running or not. This replaces the
check in the volume api tests for a cinder endpoint.

Change-Id: Id894fc29ce2635f1f058290d93726e789c909a86