Add boto tests for idempotent RunInstances calls
The EC2 API supports providing a unique client token to
the RunInstances call to ensure idempotency. In other words,
multiple RunInstances call with the same client token will
not start up multiple instances.

This adds a test for this condition, including ensuring that
a new instance is spawned when a client token is re-used
after the original instance linked to it has been terminated.

Fixes bug #1202404

Change-Id: Ia11eb9ac082a623a1ad03b89fb6dbf7cad9f9c99