Add isolated tenants for admin compute tests
This adds support for isolated credentials on admin tests. Previously
the admin tests just ran with the admin username, password, and
tenant from the config file. This changes that behavior by getting
isolated credentials for the admin base test class and then adding
the create user to the admin role.

This may fix a race condition uncovered by the testr runs.

Possibly part of: bp speed-up-tempest

Change-Id: I74339662ab769e005cb79b2d741a04640c5546ce