Move isolated credential code to BaseTestCase
This commit takes the isolated credential code from BaseComputeTest
and moves it up to it's super class: BaseTestCase. This allows the
isolated credential code to be reused by all the test classes. All
existing(only compute and volume) isolated credential methods in
the base classes are then update to use the common methods.

Change-Id: Ifae7bb74a4aa8f853fbd87ad80a63a9dcfc3b9d1