Add locking to test_aggregates
There is a race condition when running testr in parallel between
the XML and JSON versions of:

test_aggregate_add_host_create_server_with_az

because a host can only be in one availability zone at a time. This
commit adds file locking to that test using oslo lockutils so that
the XML and JSON tests will never run at the same time.

part of blueprint speed-up-tempest

Change-Id: I1ce76ccd9c57d8f44f19958498c632910112528b