Switch to using Oslo logging
This commit removes the tempest.common.log module and switches to
using the Oslo log module. The Oslo log module is required by
lockutils and it's dependent modules, so this patch migrates all of
Tempest to it so that there is only one logger in the project. This
also has the advantage of bringing Tempest's logging configuration
in line with other OpenStack projects.

Change-Id: I31b460e454123a4211c0a8a1dc91bbfd562ce81a