Add wait_for_volume_status in upload_to_image test
The upload to image test was relying on the status of the
image being uploaded to indicate whether things were completed
or not before exiting.  It appears that in some cases there
is a timing condition here where the volume's status may not
have been updated from in-use back to available before the image
is ready, resulting in tearDown being called and attempting to
delete an in-use volume which raises an unhandled exception in
the tearDown.

Fixes bug: 1202254

Change-Id: If39caa3fdd301ba4b35670b103251e89f20b3edb