Add tox job for serial testr
This commit adds a new tox job to run tempest using testr serially.
This is intended to be used for development as an alternative to
using nose while we are debugging race conditions with parallel testr.
This commit also adds another wrapper shell script so that we can run
this tox job with the colorizer script and print the slowest tests at
the end.

Optionally if you'd like to run a subset of tempest (not what runs in
the gate) you can directly call tools/pretty_tox_serial.sh with the
regex filter as the arguments.

Change-Id: I6c69bc25ebaef7df1e2389ea2c7e0ef4ca5ec2aa